chore: split out DHT publishing code to a separate file
