+++ utils/string.js
@@ -1,0 +1,4 @@
+
+export function shorten (str) {
+  return `${str.slice(0, 6)}..${str.slice(-6)}`
+}
